test: update and repair tests
