--- __mocks__/styled-jsx/css.js
@@ -1,5 +1,0 @@
-function css() {
-  return ''
-}
-
-module.exports = css
