+++ fileTransformer.js
@@ -1,0 +1,8 @@
+// Lets Jest represent imported assets with their filenames
+const path = require('path')
+
+module.exports = {
+  process(src, filename, config, options) {
+    return 'module.exports = ' + JSON.stringify(path.basename(filename)) + ';'
+  },
+}
