+++ jest-preprocess.js
@@ -1,6 +1,6 @@
 const babelOptions = {
   presets: ['babel-preset-gatsby'],
-  plugins: ['styled-jsx/babel-test'],
+  plugins: ['styled-jsx/babel'],
 }
 
 module.exports = require('babel-jest').createTransformer(babelOptions)
