+++ setup-test-env.js
@@ -1,4 +1,1 @@
-import '@testing-library/jest-dom/extend-expect'
-
-// this is basically: afterEach(cleanup)
-import '@testing-library/react/cleanup-after-each'
+import '@testing-library/jest-dom'
